Add a test on predefined_request_group at unpublication

+++ src/Ign/Bundle/OGAMConfigurateurBundle/Utils/ModelUnpublication.php
@@ -664,9 +664,11 @@ class ModelUnpublication extends DatabaseUtils {
 		$sql = "DELETE FROM website.predefined_request WHERE dataset_id = '" . $datasetId . "'";
 	 	$stmt = $this->conn->prepare($sql);
 	 	$stmt->execute();
-	 	
-	 	$sql = "DELETE FROM website.predefined_request_group WHERE group_id = '" . $datasetGroupId . "'";
-	 	$stmt = $this->conn->prepare($sql);
-	 	$stmt->execute();
+
+		if ($datasetGroupId) {
+			$sql = "DELETE FROM website.predefined_request_group WHERE group_id = '" . $datasetGroupId . "'";
+			$stmt = $this->conn->prepare($sql);
+			$stmt->execute();
+		}
 	}
 }
